# indexer/services/comlink/public/swagger.json
{
  "components": {
...
    "schemas": {
...
      "PositionSide": {
        "enum": [
          "LONG",
          "SHORT"
        ],
        "type": "string"
      },
      "IsoString": {
        "type": "string"
      },
...
      "AssetPositionResponseObject": {
        "properties": {
          "symbol": {
            "type": "string"
          },
          "side": {
            "$ref": "#/components/schemas/PositionSide"
          },
          "size": {
            "type": "string"
          },
          "assetId": {
            "type": "string"
          },
          "subaccountNumber": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "symbol",
          "side",
          "size",
          "assetId",
          "subaccountNumber"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "AssetPositionResponse": {
        "properties": {
          "positions": {
            "items": {
              "$ref": "#/components/schemas/AssetPositionResponseObject"
            },
            "type": "array"
          }
        },
        "required": [
          "positions"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "CandleResolution": {
        "enum": [
          "1MIN",
          "5MINS",
          "15MINS",
          "30MINS",
          "1HOUR",
          "4HOURS",
          "1DAY"
        ],
        "type": "string"
      },
      "CandleResponseObject": {
        "properties": {
          "startedAt": {
            "$ref": "#/components/schemas/IsoString"
          },
          "ticker": {
            "type": "string"
          },
          "resolution": {
            "$ref": "#/components/schemas/CandleResolution"
          },
          "low": {
            "type": "string"
          },
          "high": {
            "type": "string"
          },
          "open": {
            "type": "string"
          },
          "close": {
            "type": "string"
          },
          "baseTokenVolume": {
            "type": "string"
          },
          "usdVolume": {
            "type": "string"
          },
          "trades": {
            "type": "number",
            "format": "double"
          },
          "startingOpenInterest": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "startedAt",
          "ticker",
          "resolution",
          "low",
          "high",
          "open",
          "close",
          "baseTokenVolume",
          "usdVolume",
          "trades",
          "startingOpenInterest",
          "id"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "CandleResponse": {
        "properties": {
          "candles": {
            "items": {
              "$ref": "#/components/schemas/CandleResponseObject"
            },
            "type": "array"
          }
        },
        "required": [
          "candles"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "HistoricalFundingResponseObject": {
        "properties": {
          "ticker": {
            "type": "string"
          },
          "rate": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "effectiveAt": {
            "$ref": "#/components/schemas/IsoString"
          },
          "effectiveAtHeight": {
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "rate",
          "price",
          "effectiveAt",
          "effectiveAtHeight"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "HistoricalFundingResponse": {
        "properties": {
          "historicalFunding": {
            "items": {
              "$ref": "#/components/schemas/HistoricalFundingResponseObject"
            },
            "type": "array"
          }
        },
        "required": [
          "historicalFunding"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "Klyra Indexer API",
    "version": "v1.0.0",
    "description": "Scroll down for code samples, example requests and responses.\nBase URLs:\n* For **Testnet**, use <a href=\"https://klyra-testnet.imperator.co/v4\">https://klyra-testnet.imperator.co/v4</a>\nNote: Messages on Indexer WebSocket feeds are typically more recent than data fetched via Indexer's REST API, because the latter is backed by read replicas of the databases that feed the former. Ordinarily this difference is minimal (less than a second), but it might become prolonged under load."
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/assetPositions": {
      "get": {
        "operationId": "GetAssetPositions",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetPositionResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subaccountNumber",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      }
    },
    "/assetPositions/parentSubaccountNumber": {
      "get": {
        "operationId": "GetAssetPositionsForParentSubaccount",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetPositionResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "parentSubaccountNumber",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      }
    },
    "/candles/perpetualMarkets/{ticker}": {
      "get": {
        "operationId": "GetCandles",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resolution",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CandleResolution"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "fromISO",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "toISO",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/historicalFunding/{ticker}": {
      "get": {
        "operationId": "GetHistoricalFunding",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalFundingResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "effectiveBeforeOrAtHeight",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "effectiveBeforeOrAt",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IsoString"
            }
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://klyra-testnet.imperator.co/v4",
      "description": "Testnet"
    }
  ]
}